third_party/WebKit/LayoutTests/external/wpt/web-animations/timing-model/animations/set-the-timeline-of-an-animation-expected.txt

This is a testharness.js-based test.
FAIL After setting timeline on paused animation it is still pausedCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After setting timeline on animation paused outside active interval it is still pausedCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After setting timeline on an idle animation without a start time it is still idleCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After setting timeline on an idle animation with a start time it is runningassert_equals: expected "idle" but got "running"
FAIL After setting timeline on an idle animation with a sufficiently ancient start time it is finishedassert_equals: expected "idle" but got "finished"
FAIL After setting timeline on a play-pending animation it is still pendingCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After setting timeline on a play-pending animation it begins playing after pendingCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After setting timeline on a pause-pending animation it is still pendingCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After setting timeline on a pause-pending animation it becomes paused after pendingCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After clearing timeline on paused animation it is still pausedCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After clearing timeline on finished animation it is idleCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After clearing timeline on running animation it is idleCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After clearing timeline on idle animation it is still idleCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After clearing timeline on play-pending animation it is still pendingCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After clearing and re-setting timeline on play-pending animation it begins to playCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After clearing timeline on a pause-pending animation it is still pendingCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After clearing and re-setting timeline on a pause-pending animation it becomes pausedCannot assign to read only property 'timeline' of object '#<Animation>'
FAIL After clearing and re-setting timeline on an animation in the middle of an aborted pause, it continues playing using the same start timeCannot assign to read only property 'timeline' of object '#<Animation>'
Harness: the test ran to completion.
